gpu_ctx_vk: be explicit about double to int64_t conversion for present time offset

+++ libnopegl/src/backends/vk/gpu_ctx_vk.c
@@ -756,7 +756,7 @@ static VkResult swapchain_present_buffer(struct gpu_ctx *s, double t)
          */
         if (s_priv->present_time_offset == 0)
             s_priv->present_time_offset = ngli_gettime_relative() * 1000;
-        present_time.desiredPresentTime = s_priv->present_time_offset + t * 1000000000LL;
+        present_time.desiredPresentTime = s_priv->present_time_offset + (int64_t)(t * 1000000000LL);
         present_info.pNext = &present_time_info;
     }
